Refactor: Improve probability calculations with NaN handling
The previous probability calculations in the statistics module used pandas' `mean()` method, which implicitly handles NaN values. This commit refactors these calculations to be more explicit and robust.

- Explicitly count non-NaN members for each forecast time step to use as a denominator.
- Use `np.where` to avoid division-by-zero errors when all members are NaN, returning a probability of 0.
- Update `calculate_octa_probabilities`, `calculate_wind_direction_probabilities`, and `calculate_weather_code_probabilities` with this new logic.
- Remove unnecessary `.astype(int)` conversion in `weather_model.py` when calculating cloud cover octas, as the updated statistics functions can handle floats.

+++ src/open_meteo_cast/weather_model.py
@@ -126,7 +126,7 @@ class WeatherModel:
                     self.statistics[variable] = calculate_precipitation_statistics(data_df)
                 elif variable == 'cloud_cover':
                     # Convert cloud cover from percentage to octas
-                    octas_df = (data_df / 100 * 8).round().astype(int)
+                    octas_df = (data_df / 100 * 8).round()
                     self.statistics[variable] = calculate_octa_probabilities(octas_df)
                 elif variable == 'wind_direction_10m':
                     self.statistics[variable] = calculate_wind_direction_probabilities(data_df)
